+++ SpringWebApp/src/main/java/isamrs/repository/DijagnozaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import isamrs.domain.Dijagnoza;
 
-public interface DijagnozaRepository extends JpaRepository<Dijagnoza, Integer> {
+public interface DijagnozaRepository extends JpaRepository<Dijagnoza, Long> {
 
 }
